Updated to .net 6
