+++ DataAccess/DBConnectionInfo.cs
@@ -1,0 +1,13 @@
+namespace OgreDA.DataAccess;
+
+public class DBConnectionInfo
+{
+    public string ProviderName { get; set; }
+    public string ConnectionString { get; set; }
+
+    public DBConnectionInfo(string connectionString, string provider)
+    {
+        ConnectionString = connectionString;
+        ProviderName = provider;
+    }
+}
