--- OgreDA.DataAccess/Database.cs
@@ -1,16 +1,0 @@
-﻿using System;
-using System.Data;
-using System.Data.Common;
-
-namespace OgreDA.DataAccess
-{
-	public partial class Database
-	{
-	
-		private bool handleProviderSpecificParameters(DbCommand command)
-		{
-			return false;
-		}
-
-	}
-}
