--- OgreDA/DataAccessTests.Core/appsettings.json
@@ -1,10 +1,0 @@
-{
-  "ConnectionStrings": 
-    {
-      "TestDB":       {
-        "Name": "TestDB",
-        "ProviderName": "System.Data.SqlClient",
-        "ConnectionString": "Data Source=(local)\\SQLExpress;Initial Catalog=OgreDATest;Persist Security Info=True;User ID=TestUser;Password=testusrPWD!"
-      }
-    }
-}
