--- OgreDA/DataAccessTests/App.config
@@ -1,12 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<configuration>
-	<connectionStrings>
-		<add name="TestDB" providerName="System.Data.SqlClient"
-      connectionString="Data Source=(local)\SQLExpress;Initial Catalog=OgreDATest;Persist Security Info=True;User ID=TestUser;Password=testusrPWD!"/>
-		<add name="TestDBOleDb" providerName="System.Data.OleDb"
-      connectionString="Provider=SQLOLEDB.1;Data Source=(local)\SQLExpress;Initial Catalog=OgreDATest;Persist Security Info=True;User ID=TestUser;Password=testusrPWD!"/>
-		<add name="TestDBOracle" providerName="System.Data.OleDb"
-      connectionString="Provider=OraOleDB.Oracle;chunksize=8000;User ID=;Password=;Data Source=TestDB;DistribTx=0"/>
-		<add name="TestDBOracleODPManaged" providerName="Oracle.ManagedDataAccess.Client" connectionString="User ID=;Password=;Data Source=TestDB;"/>
-	</connectionStrings>
-</configuration>
