--- OgreDA/OgreDA.projitems
@@ -1,20 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8"?>
-<Project xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
-  <PropertyGroup>
-    <MSBuildAllProjects>$(MSBuildAllProjects);$(MSBuildThisFileFullPath)</MSBuildAllProjects>
-    <HasSharedItems>true</HasSharedItems>
-    <SharedGUID>2e998b32-546f-4234-9189-b62fbc83910a</SharedGUID>
-  </PropertyGroup>
-  <PropertyGroup Label="Configuration">
-    <Import_RootNamespace>OgreDA.DataAccess</Import_RootNamespace>
-  </PropertyGroup>
-  <ItemGroup>
-    <Compile Include="$(MSBuildThisFileDirectory)DataAccessException.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Database.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)DatabaseAsync.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)DbDataReaderExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)Enums.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)DbCommandExtensions.cs" />
-    <Compile Include="$(MSBuildThisFileDirectory)IDatabase.cs" />
-  </ItemGroup>
-</Project>
